Add blockchain module

+++ taucoin-android/src/main/java/io/taucoin/torrent/publishing/core/model/data/UserAndTx.java
@@ -17,12 +17,12 @@ public class UserAndTx extends Tx{
 
     public long senderBalance;              // 交易发送者的余额
 
-    public UserAndTx(@NonNull String chainID, String receiverPk, long amount, long fee, long txType, String memo) {
+    public UserAndTx(@NonNull String chainID, String receiverPk, long amount, long fee, int txType, String memo) {
         super(chainID, receiverPk, amount, fee, txType, memo);
     }
 
     @Ignore
-    public UserAndTx(@NonNull String chainID, long fee, long txType, String memo) {
+    public UserAndTx(@NonNull String chainID, long fee, int txType, String memo) {
         super(chainID, fee, txType, memo);
     }
